+++ docker-compose.yml
@@ -11,7 +11,7 @@ services:
     build:
       context: .
       dockerfile: tests/config/Dockerfile
-    image: razemax_test
+    image: aiorazemax_test
     volumes:
     - .:/app
     env_file:
@@ -25,7 +25,7 @@ services:
     build:
       context: .
       dockerfile: tests/config/Dockerfile
-    image: razemax_test
+    image: aiorazemax_test
     volumes:
     - .:/app
     command: ["pytest", "-m", "not integration"]
